Create ros unit tests, add header files
- Create unit tests for class dynamic_transform and fake_scanner
- Create header files for the class dynamic_transform and fake_scanner
